Fix debug / SERVER_MODE settings
These were out of sync at the point we started tracking dev branches.
Applying patches from one SVN dev tag to the next does not see the change
in these settings, so it stayed out of date in our git repo. Should
only have to do this once as long as we track dev.