Define an IETFJSONField that allows the code to override which values are considered blank, aka empty
The JSONField object inherits empty_values from django's base Field class, which defines it as `[ None, '', {}, [], () ]`.  However, an entered JSON string of '[]' or '{}' will be interpreted as `[]` or `{}`, respectively, which will cause it to be evaluated as empty, which causes it to fail its validation for `required`, which is added when the field is defined as `blank=False`.

This commit gives the developer options to override the empty_values, either through specifying `empty_values` directly, or by specifying `acceptable_empty_values`, a list of objects to be removed from the `empty_values` array.