Leave out traceback when emailing admins about a nomcom feedback_email error
It's possible that an error message will include part of a confidential feedback
email that should not be sent to the admins. To avoid this, give only the type
and location of the error in the main body of the email. Do not attach the
traceback for nomcom feedback_email errors.