Fill in the on_agenda field for sessions requests; refactor sreq confirmation
  * Set the on_agenda for a field based on its purpose's value
  * Call session_forms.save() when confirming a new session request
    - Fill in attendees from main form via JS

Still need to fill in the comments field for sessions when confirming a request.
This uses JS on the browser to make the per-session attendees field track the
single field visible in the UI. It would be preferable to handle this purely in
Django, but ModelForm formsets don't provide an easy way to do this. Having
spent too long on it already, I'm running with this for now because it works.
The sreq app will hopefully be going away in the near future, so it will be better
to solve this properly when replacing it.