When navigating to signin from logout, set next page to /
When a user logs in, the navigation flow of the authentication logic redirects the user back to the page from which the user clicked the "Sign In" button.  This is flawed when the user logs in from the logout page, because it logs them back out again.

This change will filter the logout page path - "accounts/logout" - out of the "next" parameter for the login page, so that the user will instead be redirected to the site root.

Fixes Trac3478.